+++ src/agents/error.rs
@@ -2,11 +2,15 @@ use crate::errors::error_chain_fmt;
 use crate::language_models::error::ModelEndpointError;
 use std::fmt::{Debug, Display, Formatter, Result as FmtResult};
 
+use super::listeners;
+
+pub type AgentResult<T> = Result<T, AgentError>;
 #[derive(thiserror::Error)]
 pub enum AgentError {
     #[error(transparent)]
     Undefined(#[from] anyhow::Error),
     ModelError(#[from] ModelEndpointError),
+    Listener(#[from] listeners::error::ListenerError),
     EnvSend,
     AgentSenderIsNone,
 }