+++ src/language_models/huggingface/completions.rs
@@ -1,0 +1,6 @@
+mod tests {
+    use rust_bert::bert::BertModel;
+
+    #[test]
+    fn comp_test() {}
+}