+++ src/language_models/huggingface/mod.rs
@@ -1,2 +1,3 @@
+pub mod completions;
 pub mod sentence_embeddings;
 pub use sentence_embeddings::embed;