+++ src/language_models/mod.rs
@@ -11,8 +11,7 @@ use anyhow::anyhow;
 use reqwest_streams::JsonStreamResponse;
 
 use crate::{
-    environment::agent_handle::MessageStack,
-    language_models::openai::completions::streaming::StreamResponse,
+    agents::memory::MessageStack, language_models::openai::completions::streaming::StreamResponse,
 };
 use reqwest::Client;
 use serde::{Deserialize, Serialize};
@@ -27,7 +26,9 @@ use self::{
         LLMEmbeddingHandler, LLMInferenceHandler,
     },
     openai::completions::{
-        functions::Function, streaming::CompletionStream, OpenAiCompletionHandler,
+        functions::Function,
+        streaming::{CompletionStream, StreamedCompletionHandler},
+        OpenAiCompletionHandler,
     },
 };
 
@@ -42,6 +43,9 @@ pub struct LLM {
     handler: LLMInferenceHandler,
     /// Temperature is computed to a number between 0 and 1 by dividing this value by 100
     params: ModelParameters,
+    api_key: String,
+    #[serde(skip)]
+    client: Client,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -96,49 +100,62 @@ impl LLM {
     pub fn new_completion_model(
         handler: LLMCompletionHandler,
         params: Option<ModelParameters>,
+        api_key: &str,
     ) -> LLM {
         let params = match params {
             None => ModelParameters::default(),
             Some(p) => p,
         };
+        let client = Client::new();
         let handler = handler.into();
-        Self { handler, params }
+        Self {
+            handler,
+            params,
+            client,
+            api_key: api_key.to_owned(),
+        }
     }
 
     pub fn new_embedding_model(
         handler: LLMEmbeddingHandler,
         params: Option<ModelParameters>,
+        api_key: &str,
     ) -> LLM {
         let params = match params {
             None => ModelParameters::default(),
             Some(p) => p,
         };
+        let client = Client::new();
         let handler = handler.into();
-        Self { handler, params }
+        Self {
+            handler,
+            params,
+            client,
+            api_key: api_key.to_owned(),
+        }
     }
 
     /// Default openai handler with 0.7 temp
-    pub fn default_openai() -> LLM {
+    pub fn default_openai(api_key: &str) -> LLM {
         let handler = OpenAiCompletionHandler::default().into();
-        LLM::new_completion_model(handler, None)
+        LLM::new_completion_model(handler, None, api_key)
     }
 
     /// Default anthropic handler with 0.7 temp
-    pub fn default_anthropic() -> LLM {
+    pub fn default_anthropic(api_key: &str) -> LLM {
         let handler = AnthropicCompletionHandler::default().into();
-        LLM::new_completion_model(handler, None)
+        LLM::new_completion_model(handler, None, api_key)
     }
 
     pub async fn get_io_completion(
         &self,
         messages: &MessageStack,
-        api_key: &str,
-        client: &Client,
     ) -> Result<String, ModelEndpointError> {
         let comp_handler: Box<dyn CompletionEndpointHandler> = self.handler.comp_handler()?;
-        let headers = comp_handler.request_headers(api_key);
+        let headers = comp_handler.request_headers(&self.api_key);
         let body = comp_handler.io_request_body(messages, self.params.temperature()?);
-        let response = client
+        let response = self
+            .client
             .post(comp_handler.completion_url())
             .headers(headers)
             .json(&body)
@@ -151,41 +168,43 @@ impl LLM {
     pub async fn get_stream_completion(
         &self,
         messages: &MessageStack,
-        api_key: &str,
-        client: &Client,
-    ) -> Result<CompletionStream, ModelEndpointError> {
+    ) -> Result<StreamedCompletionHandler, ModelEndpointError> {
         let comp_handler: Box<dyn CompletionEndpointHandler> = self.handler.comp_handler()?;
-        let headers = comp_handler.request_headers(api_key);
+        let headers = comp_handler.request_headers(&self.api_key);
         let body = comp_handler.stream_request_body(messages, self.params.temperature()?)?;
-        let request = client
+        let request = self
+            .client
             .post(comp_handler.completion_url())
             .headers(headers)
             .json(&body);
-        let response_stream = tokio::time::timeout(Duration::from_secs(10), async {
-            request
-                .send()
-                .await
-                .map_err(|err| ModelEndpointError::NetRequest(err))
-                .unwrap()
-                .json_array_stream::<StreamResponse>(1024)
-        })
-        .await
-        .map_err(|_| ModelEndpointError::Undefined(anyhow!("Response stream request timed out")))?;
+        let response_stream: CompletionStream = Box::new(
+            tokio::time::timeout(Duration::from_secs(10), async {
+                request
+                    .send()
+                    .await
+                    .map_err(|err| ModelEndpointError::NetRequest(err))
+                    .unwrap()
+                    .json_array_stream::<StreamResponse>(1024)
+            })
+            .await
+            .map_err(|_| {
+                ModelEndpointError::Undefined(anyhow!("Response stream request timed out"))
+            })?,
+        );
         tracing::info!("Got response stream, returning");
-        Ok(Box::new(response_stream))
+        Ok(response_stream.into())
     }
 
     pub async fn get_fn_completion(
         &self,
         messages: &MessageStack,
-        api_key: &str,
-        client: &Client,
         function: Function,
     ) -> Result<Value, ModelEndpointError> {
         let comp_handler: Box<dyn CompletionEndpointHandler> = self.handler.comp_handler()?;
-        let headers = comp_handler.request_headers(api_key);
+        let headers = comp_handler.request_headers(&self.api_key);
         let body = comp_handler.fn_request_body(messages, function)?;
-        let response = client
+        let response = self
+            .client
             .post(comp_handler.completion_url())
             .headers(headers)
             .json(&body)
@@ -195,16 +214,12 @@ impl LLM {
         Ok(comp_handler.handle_fn_response(json)?)
     }
 
-    pub async fn get_embedding(
-        &self,
-        text: &str,
-        api_key: &str,
-        client: &Client,
-    ) -> Result<Vec<f32>, ModelEndpointError> {
+    pub async fn get_embedding(&self, text: &str) -> Result<Vec<f32>, ModelEndpointError> {
         let emb_handler: Box<dyn EmbeddingEndpointHandler> = self.handler.emb_handler()?;
-        let headers = emb_handler.request_headers(api_key);
+        let headers = emb_handler.request_headers(&self.api_key);
         let body = emb_handler.request_body(text);
-        let response = client
+        let response = self
+            .client
             .post(emb_handler.completion_url())
             .headers(headers)
             .json(&body)