+++ tests/api/main.rs
@@ -1,4 +1,5 @@
-pub mod environment;
+pub mod agent;
+// pub mod environment;
 
 #[cfg(any(feature = "embedding_sql", feature = "tools"))]
 pub mod features;